Move src/DataViewer -> src/Database/DataViewer
