+++ src/components/Database/DataViewer/NodeLeaf.scss
