+++ src/components/Database/DataViewer/NodeParent.scss
