+++ src/components/Database/DataViewer/NodeTabularDisplay.scss
