+++ src/components/Database/DataViewer/ValueDisplay.scss
