+++ src/components/Database/DataViewer/common/_mixins.scss
