+++ src/components/Database/DataViewer/index.scss
@@ -1,5 +1,5 @@
 @import 'common/color';
-@import '../../../node_modules/@rmwc/data-table/data-table.css';
+@import '../../../../node_modules/@rmwc/data-table/data-table.css';
 
 .material-icons:hover {
   color: $COLOR_TEXT_PRIMARY;
